Create a reporting_triangle object from a data.frame

This method takes a data.frame containing case counts indexed by reference date and report date and creates a A reporting_triangle object. See as_reporting_triangle.matrix for other data input options.

Usage

# S3 method for class 'data.frame'
as_reporting_triangle(
  data,
  max_delay,
  strata = NULL,
  delays_unit = "days",
  reference_date = "reference_date",
  report_date = "report_date",
  count = "count",
  ...
)

Arguments

data

Data.frame in a long tidy format with counts by reference date and report date. Must contain the following columns: . - Column of type date or character with the dates of the primary event occurrence (reference date).

Column of type date or character with the dates of report of the primary event (report_date).

Column of numeric or integer indicating the new confirmed counts pertaining to that reference and report date (count). Additional columns can be included but will not be used. The input dataframe for this function must contain only a single strata, there can be no repeated reference dates and report dates.

max_delay

Integer indicating the maximum delay.

strata

Character string indicating the metadata on the strata of this reporting triangle. Default is NULL.

delays_unit

Character string specifying the temporal granularity of the delays. Options are "days", "weeks", "months", "years". For the matrix method, this is simply passed as an item in the reporting_triangle object which will later be used to create a nowcast data.frame. For the data.frame method, this is used to compute the delay in terms of the specified unit, and to expand the combinations of reference dates and delays to the complete set of combinations in the reporting triangle. Default is "days".

reference_date

Character string indicating the name of the column which represents the reference date, or the date of the primary event occurrence.

report_date

Character string indicating the name of the column which represents the date the primary event was reported.

count

Character string indicating the name of the column containing the number of incident cases on each reference and report date.

...

Additional arguments not used.

Value

A reporting_triangle object

See also

Examples

data_as_of_df <- syn_nssp_df[syn_nssp_df$report_date <= "2026-04-01", ]
as_reporting_triangle(
  data = data_as_of_df,
  max_delay = 25
)
#> $reporting_triangle_matrix
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#>   [1,]  194   54   26   13   12   13    5    0    0     0    14     9
#>   [2,]  198   53   29   40   11    5    1    0    0    20     0    22
#>   [3,]  167   58   33   14    0    7    4    8   12     8     0     0
#>   [4,]  145   69   22   11    5   19   23   13    8    15     0     0
#>   [5,]  180  100    0   25   12   15   31    7    4    12     5    12
#>   [6,]  220   48    0   21   11   16   24   14    0     0     0     0
#>   [7,]  326   94   14   33   42   26   27    0    0    20    11     0
#>   [8,]  202   87    0   33   31   18    0    0   26    12     0     0
#>   [9,]  256   75   13   39   19   22    7   11    5     5     5     6
#>  [10,]  213   43    1   33   14   10   23    8    3    12     4     0
#>  [11,]  229  128   28   23   10   26    8   15   31    15     8     0
#>  [12,]  237   38   24   58    1   16   15   11    0    12     0    10
#>  [13,]  242   87    8    0   13   22   15    7    0     0     0     0
#>  [14,]  325   78   36   14   26   10   16   15    7    11    26     6
#>  [15,]  211   72   29   27   25    6   10   16   27    26     0    16
#>  [16,]  340  162   35   28   22   13   11   12   23    22     0     8
#>  [17,]  241  139   14   18   14    8   16   14   25    10     1     3
#>  [18,]  258   98   26   10   13   27    8   33   15     8    21     0
#>  [19,]  279  126   23    8   35    6   33   31   13     0    12     0
#>  [20,]  176   15   11   13   10    9   32    1   19    16     0     8
#>  [21,]  167   41    2   15    0   12   12    8    0     5    12     0
#>  [22,]  246   35   20   15   11   14    9    0   14     6    18     7
#>  [23,]  210  134   53    9   22    3   16   34   22    18    12    10
#>  [24,]  157   83   20    6   16    0   30   22    7     4    13     0
#>  [25,]  202   65   30   19    0   31   12   37    4     1    24     1
#>  [26,]  239   61   21    0    8   20    0   32   13     0     0    19
#>  [27,]  160  109   25   73   32   50   40   26    0    17    11    15
#>  [28,]  310   63   19   15   26   10   18   10   17    11     8    27
#>  [29,]  268  114   32   52   27   10   17    6    8    21     5    10
#>  [30,]  266   80   42   49    6   20    6   30   22    12     6    31
#>  [31,]  287   95   41   32    8   11   11    1   18    12    12    11
#>  [32,]  202   36    7    0   17   22   21    8    0     2     0    10
#>  [33,]  288  146   28    5   41   32   28   12   17    11     0     9
#>  [34,]  223   60   26    5   14    1   11   27   11     0     5    13
#>  [35,]  246   35   11   15   21   27   22   10    0     3     0     0
#>  [36,]  273   22   20   29   35   38    0    0   11     9    17     5
#>  [37,]  289  113   10   37   28   15    0    6   27    27    28    45
#>  [38,]  305   89   21   17   13   13   33   30   21    34    17    10
#>  [39,]  284   92   18   23    9   28   32   32   14    24    10     0
#>  [40,]  279   91   16   14   12   13   31   13    0     0     6    19
#>  [41,]  225   82    8   49   27   17   28    8    3     0     0     9
#>  [42,]  174   60   26   14    7    9   26    7   13    19    15     1
#>  [43,]  286   90   32   25   21   15    0   11    2     3    26     5
#>  [44,]  190  107   45   33   34    6   19   14   10    26    31     6
#>  [45,]  291   72   23   43   11    3   56   48    8    21    14     0
#>  [46,]  323  168   36   26   38   23   50   52   45    41     0     3
#>  [47,]  192   83   11   37   33   26   21   15   10     0     0     1
#>  [48,]  151   70   44   22   38   48   14   17    8     0    13    15
#>  [49,]  213   93   26   21   15   14   14    5    1    14     7    21
#>  [50,]  187   77   28   20    8   10   13    0    6    13    14     9
#>  [51,]  272   56   58   24   36   14    5    9   14    10     8     0
#>  [52,]  205   98   25   23   29    9   25   15    1    15     0     0
#>  [53,]  268   53   47   38   16   10   10    9   12    31     0     0
#>  [54,]  177   96   23   17    8    0   14   13   14     0     0     0
#>  [55,]  213   63   34   15   23   23    0   21    8     0    11    12
#>  [56,]  218  113   25   36   26   24   10   18   16     7    22    28
#>  [57,]  135  101   17   39    9   12   10    8    7     6     4     0
#>  [58,]  161   35   17   25   12    4    7    0    9    23     7     0
#>  [59,]  256   70   14   19   28   25   32   15    9    32    12     3
#>  [60,]  160   75   46   37   17   28   25    5    9     4     0    16
#>  [61,]  207  105    5   12    8   16   25   11   11     1     0    11
#>  [62,]  233   71   22   17    8   18   18   23    1     0    12    11
#>  [63,]  194  128   35   20   19   41   17   21   11    27     0     8
#>  [64,]  233   34   23   15   22    9   12    0   19     0     0    11
#>  [65,]  163   71   39   26   22    8    5    0   20    23    14     0
#>  [66,]  137  142   22   40   15    0   10    0   18    21     0     0
#>  [67,]  210   70   36   21   18   25   11    6   13     0     0     0
#>  [68,]  127   46   21   12   25   17    0   12    0     0     0     0
#>  [69,]  172   61    6   24   16   18   11    0    0     0     0     2
#>  [70,]  119   29   25   15    0    7    0    0    7     6    13    11
#>  [71,]   93   70   29   12    8   26   10    8    6    26     0     3
#>  [72,]  172   63   32   13   20    0    0   18    0     7    19    12
#>  [73,]  126   44   13   19    0   10   38   22    0    17     5     0
#>  [74,]  265   79   31   22   16   12   44   36    7    20    13     7
#>  [75,]  168   48   16   21    6   13   23   17    1     0     2     1
#>  [76,]  176   23    9   14    0    5   12    9    0     0     0     7
#>  [77,]  187   56    8    9    0   41    3    0    8     1     0    10
#>  [78,]  133    5   18   16   28   18    0   12    0     0     8     0
#>  [79,]   96   46   20   22    8    0    9   13    1     0     0     7
#>  [80,]  194  139   56   26   13    4   13   20    0     0    14     0
#>  [81,]  234   76   30   16   12   26   12   15   10    13     7     1
#>  [82,]  225   49   24    7   31   25   18    0    7     0     4     0
#>  [83,]  105   56   10   33   22    8   14    8   10     5    11     8
#>  [84,]  148   27   29   22   15   18   15    0    0    11    13     0
#>  [85,]  162   60   34    9    0   30   10   12    6    20    11     0
#>  [86,]  207   54   55    0   23   13    7    0   11     7     6     9
#>  [87,]  239   69   41   13   12    4   35   34    1    21     8     1
#>  [88,]  189  108   22   24   24    8   17    7   20    22     0    10
#>  [89,]  176   53    9   26   12   11   34   31   12     0     0     0
#>  [90,]  276   25   38   13   42   18   30    9    0     0     7    19
#>  [91,]  260   25   29   11   12   13   20    0    0     7    18     0
#>  [92,]  203   47   16   51   29   17   13   34   26    23    20     1
#>  [93,]  330  147   37   55   13    0   13   30   27    33    28    19
#>  [94,]  331   93   37   28   40   27   37   30   41    14    26     0
#>  [95,]  188  105   27   21   26   23   23   31    9    28     0     0
#>  [96,]  225   72   16   15   26   10   47    4   23     1     0    10
#>  [97,]  275  230   36   46   30   32   30   13    7     0    36    16
#>  [98,]  197  127   35   12   14   16    9    0    3    34    22    17
#>  [99,]  242   87   17   13   20    7   16   13   17    18     8    18
#> [100,]  294  132   25   19   33   24    7   21    6    24     6     0
#> [101,]  252  134   66   71   22   11   22   16    8    24    18     9
#> [102,]  289  105   45   27   13   14   19   30    1     5     0     0
#> [103,]  297   80   85    0   44   16   27   21   27     9     0    20
#> [104,]  237   80   40   46    8   24   48   15    0    13    14     5
#> [105,]  328  145   65   47    1   45   16   27    1    16    42    20
#> [106,]  342   88   34   10   26    8   27    3    0     3    21    23
#> [107,]  444   95   48   32   33    6    0   23   50    18     0     0
#> [108,]  298  225   66   53    7   32   42   25   21    35    23     0
#> [109,]  314  137   19   19   21   34   32   31   24     9     0     1
#> [110,]  344  133   30    8   30   28   16   45   41    18    12     0
#> [111,]  402  228   56   26   34   31   35   19    0     4     6    25
#> [112,]  305   71   38   24    9   34   34    0   14     2     9     8
#> [113,]  344   89   43   32   36   13   12   20    0    10     9     0
#> [114,]  564  122   74   33   48   15   17   21    9    28    34    45
#> [115,]  417  180   51   30   12   21   11   32   23    14    14    14
#> [116,]  345  126   90   67   26   12   29   42   17     9     9    11
#> [117,]  412   88    7    1   18   16   67   28   23     0    13     0
#> [118,]  306  158   20   22   27   34   45   16   11     0     0     6
#> [119,]  406  204   42   34   34   20   29    6    6     2    15     1
#> [120,]  404  151   73   35   34   39   17    9    0    24    15    25
#> [121,]  313  178   47   12   39   30    9   21   17     9    27     1
#> [122,]  615  270   35   81   45   26   63   65   42    51    38    15
#> [123,]  470  243   73   30   35   28   34   25   22     5     0     0
#> [124,]  662  275   50   23   40   13   43   64   13     0     5    20
#> [125,]  534  159   74   76   21   52   22   50    0    10    40    22
#> [126,]  638  196   91   66   38   43   62   20   19    43    39    12
#> [127,]  591  230   71   68   58   36   12   28   30    44    28    21
#> [128,]  556  184   47   48   51   40   41   34    5    20    13    22
#> [129,]  487  252   82   69   23   13   57   33   48    49    25    23
#> [130,]  603  261   96   47   27   58   60   35   43    33    18     7
#> [131,]  411  178   65   16   27   17   23   79   46     0     8     9
#> [132,]  504  224   39   73   50   13   24   17   13    23    22     5
#> [133,]  533   96   24   45   29   36   31   41   15    14    13    12
#> [134,]  310  126   43   33   36   12   12   24   28    24    15    32
#> [135,]  414  215   56   84   46   23   19   24   26    16    35    27
#> [136,]  689  333  101   90   58   47   17   26   64    37    22     7
#> [137,]  477  102   67   21    7   30   20   31   51    11    15    11
#> [138,]  493  185   49   17   44   34   32   57   31     5     6    21
#> [139,]  340  116   69   17   17   31   26   15   12     5    32     0
#> [140,]  251  100   32   42   18   20   37   11   11     0    28     0
#> [141,]  300   99   47   34    5   45   18    0   20    16    20    14
#> [142,]  378   77   22   16   16    1   11   14   38    14     8    13
#> [143,]  508  147   37   37   13   15   32   27   16     6    11     4
#> [144,]  482  172   82    2   35   35   13   30   22     0     0     0
#> [145,]  295  197   53   47   19   18   36    8   10     0     0    20
#> [146,]  274   97   28   31   29    5    9   31    7     5     6    18
#> [147,]  171   36   34   26   13   46   11   12   18     1     8     8
#> [148,]  280  155   83   46   15   22    3    4   25     0    10    11
#> [149,]  246   95   57   19   38   14   10    0    0    10    18    NA
#> [150,]  210  131   34   50   35   12    1   25   20     6    NA    NA
#> [151,]  221   96   22   10   13    6    0    5    9    NA    NA    NA
#> [152,]  291  129   17   26   42   29   23   25   NA    NA    NA    NA
#> [153,]  179   96   22   50    9    8   18   NA   NA    NA    NA    NA
#> [154,]  284   40   41   54   28   12   NA   NA   NA    NA    NA    NA
#> [155,]  217   78   46   14   39   NA   NA   NA   NA    NA    NA    NA
#> [156,]  336  161   62   13   NA   NA   NA   NA   NA    NA    NA    NA
#> [157,]  296   53   55   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> [158,]  210  108   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> [159,]  236   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>        [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22]
#>   [1,]    14     0     0     0     6     0     0     0     0     0
#>   [2,]     0    10     7     0     0    20     0     0     0     0
#>   [3,]     6     0    19     6     9     0     0     0     0     0
#>   [4,]     0     3     0    10     0     9     0     0     0     0
#>   [5,]     5     0     8     0     0     0     5     5     0     0
#>   [6,]     8     0     0     0     0     0     0     0     0     0
#>   [7,]     2     5    14     0     5     8     2     0     0     0
#>   [8,]     0     0     0     0     5     0     0     0     0     0
#>   [9,]     4     0     0     8    12     5     0     0     0    10
#>  [10,]     8     5    22     0     0     0     0     0    11    15
#>  [11,]    13     0    11     5     0    14     0     0     0    17
#>  [12,]     0     0    13     0     1     0     4    19     9     0
#>  [13,]     0     1     0     0     0    10     4     0     2     0
#>  [14,]    15     0     0     0     1     0     0    13     0     0
#>  [15,]     0     0     0     0     9     0     7    10     0     0
#>  [16,]     9     9     0     0     0     0     0     0     3    14
#>  [17,]     0    18     0     0     1    14     3     5     8     0
#>  [18,]    32     8     7    12    24     0     0     4     0     0
#>  [19,]     0    14     0     7    20     0    17     8     9     0
#>  [20,]     7     8     4     1     0     0     0     0     0     0
#>  [21,]     1    15     0     0     0     0     0     0     3     0
#>  [22,]    18    12     0    18    12     0     0     0     0     0
#>  [23,]    10     0     0     8     0     0     0     0     0     0
#>  [24,]     0    21     0     0     8     0     0     0     0     7
#>  [25,]     0    11     0     4     0     0     0     0     0     0
#>  [26,]    13     7     1    10     0     0     0    10     0     0
#>  [27,]    19     0     0     0     0    11    12     1    11     0
#>  [28,]     1     1     0     0     0     0     1     0     0     0
#>  [29,]     0     9    17     0     5     1     0     4     0     0
#>  [30,]     6     9     0    14    21     9     0     0     0     0
#>  [31,]     0     6     8     0     7     0     0     0     0     0
#>  [32,]    16    15     1    11     6     0     7     5     0     9
#>  [33,]     8    17     0     7     9    16     8     6     0     4
#>  [34,]     7     0    18     0     0    11     0     0     0     0
#>  [35,]     0     0     0     0     0     4    13     0     0     0
#>  [36,]     0     0     8     1     0     9     0     0     0     0
#>  [37,]     0     2     9     0     1     1    13     0     0     7
#>  [38,]     0     1    13     0     0    13     0     4     0     0
#>  [39,]     8     7     0     0     0     0     0     0     7     0
#>  [40,]    17     5    14     6     0     0    10     4    15     0
#>  [41,]    24     3     4    12     0     8     0     0     0     0
#>  [42,]     0    10     0     0     7     0     0     7     8     0
#>  [43,]     1     0     0     0     0     7     0     0     0     0
#>  [44,]     5     0     5     0     0     7     0     0     0     0
#>  [45,]    13    10    12     7     8     0     0     8     0     0
#>  [46,]    16     6     6     0     0     4     0    11     8     0
#>  [47,]     3     1     0    20     9     0     0     8     0    10
#>  [48,]     1    11     0     8     0    14    15     0     8     6
#>  [49,]     3    10     0     0     0     0     0     0     0     0
#>  [50,]     0     0     0     0     0     6     0     6     0     0
#>  [51,]     0    14     6     0     0     0     0     0     0     0
#>  [52,]     9    15     0    11    14     0     0     7     7     0
#>  [53,]     7    10    10     0     0     0     0     0     0     0
#>  [54,]     5     0     6     0     0     0     0     0     0     0
#>  [55,]     1    24     0     0     0     0     0     0     1     0
#>  [56,]    10     0     4     0    13     8     1     0    10     5
#>  [57,]     0     0     0    17    14     6     0     0     4     9
#>  [58,]     7     0     9     1     0     0     0     0     0    11
#>  [59,]     3     1    10     0     0     0    13     0     0     0
#>  [60,]    12    13     0     7     4     0     0     0     0     0
#>  [61,]     0     0     8     0     0     0     8     1     9     8
#>  [62,]     0    12     7     0     0     0     8     0     0     0
#>  [63,]     8     0     0     2     0     8     0     6     0     0
#>  [64,]     0     0     0     4     0     0     4     0     0     0
#>  [65,]     0     0     0     7     0     0     0     0     0     0
#>  [66,]     0     0     0     0     0     0     0     0     0     0
#>  [67,]    24     6     0     5     0     0     0     0     0     3
#>  [68,]     7    12     0     0     0     0     0     0     0     0
#>  [69,]     5    30    16     0     0     0     0    10     0     0
#>  [70,]     0     0     0     0     0     0     5     0     0     0
#>  [71,]     0    14     0    16    14     0     0     0     0     0
#>  [72,]     0     0     0     4     0     0     0     0     0     1
#>  [73,]     0     0     5     0     0     0     0     0     0     0
#>  [74,]    10     6     6     0     8     0     0     8     0     5
#>  [75,]    27     0     0     0     0     0     0     0     0     0
#>  [76,]     0     0     0     0     0     0     0     0     0     0
#>  [77,]     7     1     0    15     3     0    17     0     0     0
#>  [78,]     0     0     0     0     0     0     0     7     0     0
#>  [79,]     0     0    15     0     0    17     0     0     0     0
#>  [80,]     6    16     6    13     1     0    11     0     0     0
#>  [81,]     5     0    15    11     0    13     0     0     7     5
#>  [82,]    15     0     0    14     0     0     5    11     0     0
#>  [83,]     8     9     9     5     0    10     0     0     0     0
#>  [84,]     0     1     0     0     0     0     0     0     0     0
#>  [85,]    12     0     0     7     8     3     6     0     0     2
#>  [86,]     0     0    12    10    11     0     0     0     0     0
#>  [87,]     0    15    36     0     0    19     5     9     0     6
#>  [88,]    15    19    10     2     6     0     0     0     0     0
#>  [89,]     1    11     8     0     0     0     0     0    14     6
#>  [90,]     2     8     0     0     6    18     0     8     0     0
#>  [91,]     4     5    11     0     0     0     0     0     0     0
#>  [92,]    10     6     0     0     0     3     0     6    14     4
#>  [93,]     0     0     0     0     0    13     0     0     0     0
#>  [94,]     7     6     0    10     4     0     0     0     0     0
#>  [95,]    15    12    13     1     2     0     1    14     0     0
#>  [96,]     9    16     0     0     0     0     0     1    14     0
#>  [97,]     6     0    17     0     0     2     0     5     9     0
#>  [98,]    16     0     9     0     1     0     0    11     0    10
#>  [99,]     3     0     0     0     9     0    10    10     0     0
#> [100,]     0     0    16     5     0     0     4     0     0     0
#> [101,]     0     0    16     7     9    21     0     0     0    12
#> [102,]    33     7    12     9    10     0     0     0     8     0
#> [103,]    17     5    18     0     0     0     1     8     1    13
#> [104,]    16    17     3     0     0     0     1     8     0    13
#> [105,]    30    19     0     0     0    15     1     0     0     0
#> [106,]     4     7     0     0     9     5     5     0     0     5
#> [107,]     8     0     0    13    21     0    11     0     0     6
#> [108,]    15    16     1     5    21     0     0     0     0     0
#> [109,]    10    14    15    17     0     0     0     0     3     7
#> [110,]    13    23     3    25    15     0     0     0    21     0
#> [111,]    15    12     3    10     0     3    11    20    17    17
#> [112,]    44    13    12     0     6     0     7     0     8     0
#> [113,]    43     8     3     3    29    21     9     1     0     0
#> [114,]     0     0    15    32    10     0     0    12     0     5
#> [115,]     0     0    25    32    18    14    14     0     7     8
#> [116,]     0    26    19     5     8     0     6    32     5     5
#> [117,]    20    12    29    28     0     0    13    21     0     0
#> [118,]    28    37    11     6     0     0     0    14     3     6
#> [119,]    11    46     0    16     8     9     9     0     5     0
#> [120,]    30     0     0     9    26     3     2    13     0     0
#> [121,]     8     0     0    13    25    16    29     0     0    10
#> [122,]    14     6    10    35    38    51     0     2    10     5
#> [123,]     4    16     1    25    32     0     0     0    16     4
#> [124,]     3    14    11    17     6     0     0    22    10     6
#> [125,]     6    19     0     5     0     3     0    33    18     7
#> [126,]    14    18     0    18     0     1     3    11    15     0
#> [127,]    25     8     0    14    31    16     8     0     0     0
#> [128,]     0     8    10    19     6     8    18     0    19     0
#> [129,]     1     9    27    15    10     5    24    17     8     6
#> [130,]    30    37    20    18     7    12    13     0     7     0
#> [131,]     0    13     0     8     3     0     4     5     0     0
#> [132,]    18    28     3     5     0    17    10     0    13     6
#> [133,]    33    19     0     0    20    29    31     5     0     0
#> [134,]    15    17     7    10     2     8     1    14     0     0
#> [135,]     0     1     0     9    10     8     0    15     0     0
#> [136,]    10    16    31     0    11    15     0     0    20     6
#> [137,]    20     0    17    31    14     0     0    12     7     6
#> [138,]    26     6    14     7     0     0     9    15     0     0
#> [139,]    34     4     0     4     0    11     9     0     9    NA
#> [140,]     8     6     0     7    13     7    14    11    NA    NA
#> [141,]    19     0     0     0     0     3     0    NA    NA    NA
#> [142,]     0     0     0    16     3    16    NA    NA    NA    NA
#> [143,]     6    23     0    10    14    NA    NA    NA    NA    NA
#> [144,]     6     0     0     2    NA    NA    NA    NA    NA    NA
#> [145,]    18     2    21    NA    NA    NA    NA    NA    NA    NA
#> [146,]    14    17    NA    NA    NA    NA    NA    NA    NA    NA
#> [147,]    31    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [148,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [149,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [150,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [151,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [152,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [153,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [154,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [155,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [156,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [157,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [158,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [159,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>        [,23] [,24] [,25] [,26]
#>   [1,]    12     0    10     0
#>   [2,]     0    10     0     6
#>   [3,]     0     0     0     0
#>   [4,]     9     0     0     0
#>   [5,]     0     0     0     0
#>   [6,]     0     0     0     0
#>   [7,]     0     0     0     0
#>   [8,]     0     0     0     0
#>   [9,]    13     0     0     0
#>  [10,]     0     0     0     0
#>  [11,]     5     0     6    13
#>  [12,]     9     0     0     8
#>  [13,]     0     0     0     4
#>  [14,]     0     0     7     0
#>  [15,]     6     0    14     0
#>  [16,]     0     0     0     0
#>  [17,]     0     0     0     1
#>  [18,]     0     3     0     9
#>  [19,]     0     0     1     0
#>  [20,]     0     0     0     0
#>  [21,]     0     0     0     0
#>  [22,]     6     0     0     0
#>  [23,]     9    12     0     0
#>  [24,]    10     0     0     0
#>  [25,]    12     0     0     0
#>  [26,]     0     0     0     0
#>  [27,]     0     0     0     8
#>  [28,]     0     1     0     0
#>  [29,]     0     1     0     2
#>  [30,]     0     0     0     0
#>  [31,]     0     0     0     0
#>  [32,]     5     0     0     0
#>  [33,]     0    10     0     7
#>  [34,]     0     0     0     0
#>  [35,]     0     0     0     0
#>  [36,]     0     0     0     0
#>  [37,]     2     0     0     0
#>  [38,]     0     0    17     0
#>  [39,]     0     0     0     0
#>  [40,]     7    15     0     0
#>  [41,]     0     0     0     0
#>  [42,]     0     0     0     0
#>  [43,]     0     0     0     2
#>  [44,]     0     1    16    11
#>  [45,]     0     0     3     0
#>  [46,]     0     0    23     0
#>  [47,]     0     0     0     0
#>  [48,]     0    10     0     0
#>  [49,]     0     0    14     0
#>  [50,]     0     0     0     0
#>  [51,]    15     0     0     0
#>  [52,]     0     0     0     0
#>  [53,]     0     0     0     0
#>  [54,]     0     0     0     0
#>  [55,]     0     0     0     0
#>  [56,]     0     0    18     0
#>  [57,]     0    16     0     0
#>  [58,]     0     4     0     0
#>  [59,]     0     0    14     0
#>  [60,]     0     0     0     0
#>  [61,]     0     0     0     7
#>  [62,]     0     0     0     0
#>  [63,]     0     0     1     0
#>  [64,]     0     4     0     0
#>  [65,]     0     0     0     5
#>  [66,]     0     1     0     0
#>  [67,]     0     0     0     0
#>  [68,]     0     0     0     0
#>  [69,]     0     0     0     0
#>  [70,]     0     0     2     0
#>  [71,]     0     0     0     0
#>  [72,]     0     5     0     0
#>  [73,]     0     0     0     0
#>  [74,]     0     0     0     0
#>  [75,]     0     0     0     0
#>  [76,]     0     0     0     1
#>  [77,]     0     0     4     6
#>  [78,]     0     0     0     9
#>  [79,]     0     0     1     0
#>  [80,]     0     8     0     0
#>  [81,]     8    10     0     0
#>  [82,]     4     0     0     8
#>  [83,]     0     0     1     0
#>  [84,]     0     0     0     0
#>  [85,]     0     6     0     0
#>  [86,]     0     0     0    16
#>  [87,]     0     0     0     0
#>  [88,]     0     0     0     0
#>  [89,]     0     0     0     0
#>  [90,]     0     0     0     4
#>  [91,]     0     0     0     0
#>  [92,]     0     5     0     0
#>  [93,]     0     7     0     0
#>  [94,]     0     0     9     0
#>  [95,]     0     0     0     0
#>  [96,]     6     0     0     0
#>  [97,]     0     0     0     0
#>  [98,]     8     0     0     7
#>  [99,]     0     0     0     1
#> [100,]     0     0     0     0
#> [101,]    13     0     0     4
#> [102,]     0     0     0     0
#> [103,]     7     0     0     0
#> [104,]     0     0     0     0
#> [105,]     0     0     0     0
#> [106,]     0     0     0     0
#> [107,]     6     0     0     0
#> [108,]     0     0     0     0
#> [109,]     1     0     0     1
#> [110,]     1     0     0     0
#> [111,]    14     0     7     0
#> [112,]     0     0     9     0
#> [113,]     0    13     0     0
#> [114,]     0    20     0     0
#> [115,]     9     0     0    13
#> [116,]    12     3     0     0
#> [117,]     0     0     4     9
#> [118,]     0     3     0    19
#> [119,]     0     0     0    10
#> [120,]     0     0     0    20
#> [121,]     0    13     9    42
#> [122,]    23     3    11    14
#> [123,]     0     0     0     0
#> [124,]    10     2     1    11
#> [125,]     0     1     6     9
#> [126,]     9     3     0     0
#> [127,]     0     9    13     0
#> [128,]     7     0     0    12
#> [129,]     5     4    14     6
#> [130,]    13     3     0     0
#> [131,]    15     9     0     9
#> [132,]     8     0     0     0
#> [133,]     0     0    16    10
#> [134,]     7     0     0     5
#> [135,]    10     0     4    NA
#> [136,]     3    12    NA    NA
#> [137,]    15    NA    NA    NA
#> [138,]    NA    NA    NA    NA
#> [139,]    NA    NA    NA    NA
#> [140,]    NA    NA    NA    NA
#> [141,]    NA    NA    NA    NA
#> [142,]    NA    NA    NA    NA
#> [143,]    NA    NA    NA    NA
#> [144,]    NA    NA    NA    NA
#> [145,]    NA    NA    NA    NA
#> [146,]    NA    NA    NA    NA
#> [147,]    NA    NA    NA    NA
#> [148,]    NA    NA    NA    NA
#> [149,]    NA    NA    NA    NA
#> [150,]    NA    NA    NA    NA
#> [151,]    NA    NA    NA    NA
#> [152,]    NA    NA    NA    NA
#> [153,]    NA    NA    NA    NA
#> [154,]    NA    NA    NA    NA
#> [155,]    NA    NA    NA    NA
#> [156,]    NA    NA    NA    NA
#> [157,]    NA    NA    NA    NA
#> [158,]    NA    NA    NA    NA
#> [159,]    NA    NA    NA    NA
#> 
#> $reference_dates
#>   [1] "2025-10-25" "2025-10-26" "2025-10-27" "2025-10-28" "2025-10-29"
#>   [6] "2025-10-30" "2025-10-31" "2025-11-01" "2025-11-02" "2025-11-03"
#>  [11] "2025-11-04" "2025-11-05" "2025-11-06" "2025-11-07" "2025-11-08"
#>  [16] "2025-11-09" "2025-11-10" "2025-11-11" "2025-11-12" "2025-11-13"
#>  [21] "2025-11-14" "2025-11-15" "2025-11-16" "2025-11-17" "2025-11-18"
#>  [26] "2025-11-19" "2025-11-20" "2025-11-21" "2025-11-22" "2025-11-23"
#>  [31] "2025-11-24" "2025-11-25" "2025-11-26" "2025-11-27" "2025-11-28"
#>  [36] "2025-11-29" "2025-11-30" "2025-12-01" "2025-12-02" "2025-12-03"
#>  [41] "2025-12-04" "2025-12-05" "2025-12-06" "2025-12-07" "2025-12-08"
#>  [46] "2025-12-09" "2025-12-10" "2025-12-11" "2025-12-12" "2025-12-13"
#>  [51] "2025-12-14" "2025-12-15" "2025-12-16" "2025-12-17" "2025-12-18"
#>  [56] "2025-12-19" "2025-12-20" "2025-12-21" "2025-12-22" "2025-12-23"
#>  [61] "2025-12-24" "2025-12-25" "2025-12-26" "2025-12-27" "2025-12-28"
#>  [66] "2025-12-29" "2025-12-30" "2025-12-31" "2026-01-01" "2026-01-02"
#>  [71] "2026-01-03" "2026-01-04" "2026-01-05" "2026-01-06" "2026-01-07"
#>  [76] "2026-01-08" "2026-01-09" "2026-01-10" "2026-01-11" "2026-01-12"
#>  [81] "2026-01-13" "2026-01-14" "2026-01-15" "2026-01-16" "2026-01-17"
#>  [86] "2026-01-18" "2026-01-19" "2026-01-20" "2026-01-21" "2026-01-22"
#>  [91] "2026-01-23" "2026-01-24" "2026-01-25" "2026-01-26" "2026-01-27"
#>  [96] "2026-01-28" "2026-01-29" "2026-01-30" "2026-01-31" "2026-02-01"
#> [101] "2026-02-02" "2026-02-03" "2026-02-04" "2026-02-05" "2026-02-06"
#> [106] "2026-02-07" "2026-02-08" "2026-02-09" "2026-02-10" "2026-02-11"
#> [111] "2026-02-12" "2026-02-13" "2026-02-14" "2026-02-15" "2026-02-16"
#> [116] "2026-02-17" "2026-02-18" "2026-02-19" "2026-02-20" "2026-02-21"
#> [121] "2026-02-22" "2026-02-23" "2026-02-24" "2026-02-25" "2026-02-26"
#> [126] "2026-02-27" "2026-02-28" "2026-03-01" "2026-03-02" "2026-03-03"
#> [131] "2026-03-04" "2026-03-05" "2026-03-06" "2026-03-07" "2026-03-08"
#> [136] "2026-03-09" "2026-03-10" "2026-03-11" "2026-03-12" "2026-03-13"
#> [141] "2026-03-14" "2026-03-15" "2026-03-16" "2026-03-17" "2026-03-18"
#> [146] "2026-03-19" "2026-03-20" "2026-03-21" "2026-03-22" "2026-03-23"
#> [151] "2026-03-24" "2026-03-25" "2026-03-26" "2026-03-27" "2026-03-28"
#> [156] "2026-03-29" "2026-03-30" "2026-03-31" "2026-04-01"
#> 
#> $structure
#> [1] 1
#> 
#> $max_delay
#> [1] 25
#> 
#> $delays_unit
#> [1] "days"
#> 
#> $strata
#> NULL
#> 
#> attr(,"class")
#> [1] "reporting_triangle"