Calculate the updates rows of the expected nowcasted triangle

Calculate the updates rows of the expected nowcasted triangle

Usage

.calc_expectation(co, expectation, n_dates, delay_pmf)

Arguments

co

Integer indicating the column index

expectation

Matrix of the partially complete reporting triangle

n_dates

Integer indicating the number of dates in the reporting triangle (number of rows in the reporting triangle)

delay_pmf

Vector specifying the probability of a case being reported with delay d

Value

Matrix with another set of entries corresponding to the updated values for the specified rows and column