Validate triangle Various checks to make sure that the reporting triangle passed in to get_delay_estimate() is formatted properly.

Validate triangle Various checks to make sure that the reporting triangle passed in to get_delay_estimate() is formatted properly.

Usage

.validate_triangle(
  triangle,
  max_delay = ncol(triangle) - 1,
  n = nrow(triangle)
)

Arguments

max_delay

Integer indicating the maximum delay to estimate, in units of the delay. The default is to use the whole reporting triangle, ncol(triangle) -1.

n

Integer indicating the number of reference times (observations) to be used in the estimate of the reporting delay, always starting from the most recent reporting delay. The default is to use the whole reporting triangle, so nrow(triangle).

Value

NULL, invisibly