Get a dataframe of multiple draws of only the predicted elements of the nowcast vector

Get a dataframe of multiple draws of only the predicted elements of the nowcast vector

Usage

get_nowcast_pred_draws(point_nowcast_pred_matrix, disp, n_draws = 1000)

Arguments

point_nowcast_pred_matrix

Matrix containing only the elements that have not yet been observed as of the final reference date.

disp

Vector of dispersion parameters indexed by horizon from minus one to the maximum delay.

n_draws

Integer indicating the number of draws of the predicted nowcast vector to generate. Default is 1000.

Value

Dataframe containing the predicted point nowcast vectors indexed by reference time (pred_count), reference time (time), and the draw index (draw).

Examples

point_nowcast_pred_matrix <-
  matrix(
    c(
      NA, NA, NA, NA,
      NA, NA, NA, NA,
      NA, NA, NA, 16.8,
      NA, NA, 21.2, 19.5,
      NA, 34.5, 15.4, 9.1
    ),
    nrow = 4,
    byrow = TRUE
  )
disp <- c(0.8, 12.4, 9.1)
nowcast_pred_draws_df <- get_nowcast_pred_draws(
  point_nowcast_pred_matrix,
  disp,
  500
)
nowcast_pred_draws_df
#>      pred_count time draw
#> 1             0    1    1
#> 2             0    2    1
#> 3            36    3    1
#> 4            11    4    1
#> 5             0    1    2
#> 6             0    2    2
#> 7            26    3    2
#> 8             4    4    2
#> 9             0    1    3
#> 10            0    2    3
#> 11           36    3    3
#> 12           47    4    3
#> 13            0    1    4
#> 14            0    2    4
#> 15           28    3    4
#> 16           12    4    4
#> 17            0    1    5
#> 18            0    2    5
#> 19           31    3    5
#> 20           62    4    5
#> 21            0    1    6
#> 22            0    2    6
#> 23           10    3    6
#> 24          144    4    6
#> 25            0    1    7
#> 26            0    2    7
#> 27           37    3    7
#> 28           25    4    7
#> 29            0    1    8
#> 30            0    2    8
#> 31           25    3    8
#> 32          109    4    8
#> 33            0    1    9
#> 34            0    2    9
#> 35           36    3    9
#> 36           38    4    9
#> 37            0    1   10
#> 38            0    2   10
#> 39           44    3   10
#> 40           27    4   10
#> 41            0    1   11
#> 42            0    2   11
#> 43           33    3   11
#> 44          109    4   11
#> 45            0    1   12
#> 46            0    2   12
#> 47           28    3   12
#> 48          192    4   12
#> 49            0    1   13
#> 50            0    2   13
#> 51           42    3   13
#> 52           66    4   13
#> 53            0    1   14
#> 54            0    2   14
#> 55           69    3   14
#> 56          109    4   14
#> 57            0    1   15
#> 58            0    2   15
#> 59           60    3   15
#> 60           90    4   15
#> 61            0    1   16
#> 62            0    2   16
#> 63           48    3   16
#> 64          100    4   16
#> 65            0    1   17
#> 66            0    2   17
#> 67           25    3   17
#> 68          160    4   17
#> 69            0    1   18
#> 70            0    2   18
#> 71           52    3   18
#> 72           85    4   18
#> 73            0    1   19
#> 74            0    2   19
#> 75           36    3   19
#> 76          134    4   19
#> 77            0    1   20
#> 78            0    2   20
#> 79           45    3   20
#> 80           95    4   20
#> 81            0    1   21
#> 82            0    2   21
#> 83           41    3   21
#> 84           62    4   21
#> 85            0    1   22
#> 86            0    2   22
#> 87           24    3   22
#> 88           10    4   22
#> 89            0    1   23
#> 90            0    2   23
#> 91           26    3   23
#> 92          268    4   23
#> 93            0    1   24
#> 94            0    2   24
#> 95           31    3   24
#> 96           23    4   24
#> 97            0    1   25
#> 98            0    2   25
#> 99           48    3   25
#> 100           8    4   25
#> 101           0    1   26
#> 102           0    2   26
#> 103          29    3   26
#> 104           8    4   26
#> 105           0    1   27
#> 106           0    2   27
#> 107          34    3   27
#> 108          42    4   27
#> 109           0    1   28
#> 110           0    2   28
#> 111          32    3   28
#> 112         266    4   28
#> 113           0    1   29
#> 114           0    2   29
#> 115          10    3   29
#> 116          75    4   29
#> 117           0    1   30
#> 118           0    2   30
#> 119          55    3   30
#> 120          11    4   30
#> 121           0    1   31
#> 122           0    2   31
#> 123          34    3   31
#> 124          76    4   31
#> 125           0    1   32
#> 126           0    2   32
#> 127          29    3   32
#> 128         319    4   32
#> 129           0    1   33
#> 130           0    2   33
#> 131          37    3   33
#> 132          26    4   33
#> 133           0    1   34
#> 134           0    2   34
#> 135          40    3   34
#> 136         170    4   34
#> 137           0    1   35
#> 138           0    2   35
#> 139          40    3   35
#> 140          15    4   35
#> 141           0    1   36
#> 142           0    2   36
#> 143          30    3   36
#> 144          46    4   36
#> 145           0    1   37
#> 146           0    2   37
#> 147          24    3   37
#> 148          67    4   37
#> 149           0    1   38
#> 150           0    2   38
#> 151          32    3   38
#> 152           4    4   38
#> 153           0    1   39
#> 154           0    2   39
#> 155          29    3   39
#> 156          18    4   39
#> 157           0    1   40
#> 158           0    2   40
#> 159          37    3   40
#> 160           0    4   40
#> 161           0    1   41
#> 162           0    2   41
#> 163          30    3   41
#> 164          39    4   41
#> 165           0    1   42
#> 166           0    2   42
#> 167          24    3   42
#> 168           1    4   42
#> 169           0    1   43
#> 170           0    2   43
#> 171          66    3   43
#> 172         196    4   43
#> 173           0    1   44
#> 174           0    2   44
#> 175          34    3   44
#> 176          15    4   44
#> 177           0    1   45
#> 178           0    2   45
#> 179          19    3   45
#> 180          19    4   45
#> 181           0    1   46
#> 182           0    2   46
#> 183          33    3   46
#> 184           3    4   46
#> 185           0    1   47
#> 186           0    2   47
#> 187          19    3   47
#> 188           7    4   47
#> 189           0    1   48
#> 190           0    2   48
#> 191          27    3   48
#> 192          76    4   48
#> 193           0    1   49
#> 194           0    2   49
#> 195          43    3   49
#> 196          97    4   49
#> 197           0    1   50
#> 198           0    2   50
#> 199          38    3   50
#> 200         141    4   50
#> 201           0    1   51
#> 202           0    2   51
#> 203          51    3   51
#> 204           0    4   51
#> 205           0    1   52
#> 206           0    2   52
#> 207          26    3   52
#> 208         142    4   52
#> 209           0    1   53
#> 210           0    2   53
#> 211          28    3   53
#> 212         344    4   53
#> 213           0    1   54
#> 214           0    2   54
#> 215          47    3   54
#> 216         186    4   54
#> 217           0    1   55
#> 218           0    2   55
#> 219          39    3   55
#> 220           2    4   55
#> 221           0    1   56
#> 222           0    2   56
#> 223          41    3   56
#> 224          46    4   56
#> 225           0    1   57
#> 226           0    2   57
#> 227          24    3   57
#> 228         215    4   57
#> 229           0    1   58
#> 230           0    2   58
#> 231          50    3   58
#> 232          22    4   58
#> 233           0    1   59
#> 234           0    2   59
#> 235          51    3   59
#> 236          70    4   59
#> 237           0    1   60
#> 238           0    2   60
#> 239          45    3   60
#> 240          30    4   60
#> 241           0    1   61
#> 242           0    2   61
#> 243          34    3   61
#> 244           3    4   61
#> 245           0    1   62
#> 246           0    2   62
#> 247          46    3   62
#> 248          65    4   62
#> 249           0    1   63
#> 250           0    2   63
#> 251          42    3   63
#> 252         103    4   63
#> 253           0    1   64
#> 254           0    2   64
#> 255          64    3   64
#> 256          33    4   64
#> 257           0    1   65
#> 258           0    2   65
#> 259          41    3   65
#> 260         125    4   65
#> 261           0    1   66
#> 262           0    2   66
#> 263          24    3   66
#> 264          23    4   66
#> 265           0    1   67
#> 266           0    2   67
#> 267          38    3   67
#> 268          74    4   67
#> 269           0    1   68
#> 270           0    2   68
#> 271          30    3   68
#> 272          85    4   68
#> 273           0    1   69
#> 274           0    2   69
#> 275          22    3   69
#> 276           8    4   69
#> 277           0    1   70
#> 278           0    2   70
#> 279          42    3   70
#> 280          25    4   70
#> 281           0    1   71
#> 282           0    2   71
#> 283          26    3   71
#> 284          22    4   71
#> 285           0    1   72
#> 286           0    2   72
#> 287          50    3   72
#> 288         213    4   72
#> 289           0    1   73
#> 290           0    2   73
#> 291          25    3   73
#> 292          19    4   73
#> 293           0    1   74
#> 294           0    2   74
#> 295          32    3   74
#> 296          43    4   74
#> 297           0    1   75
#> 298           0    2   75
#> 299          59    3   75
#> 300          85    4   75
#> 301           0    1   76
#> 302           0    2   76
#> 303          45    3   76
#> 304         227    4   76
#> 305           0    1   77
#> 306           0    2   77
#> 307          38    3   77
#> 308          13    4   77
#> 309           0    1   78
#> 310           0    2   78
#> 311          37    3   78
#> 312          22    4   78
#> 313           0    1   79
#> 314           0    2   79
#> 315          28    3   79
#> 316          93    4   79
#> 317           0    1   80
#> 318           0    2   80
#> 319          58    3   80
#> 320          43    4   80
#> 321           0    1   81
#> 322           0    2   81
#> 323          35    3   81
#> 324           9    4   81
#> 325           0    1   82
#> 326           0    2   82
#> 327          29    3   82
#> 328         335    4   82
#> 329           0    1   83
#> 330           0    2   83
#> 331          41    3   83
#> 332          15    4   83
#> 333           0    1   84
#> 334           0    2   84
#> 335          36    3   84
#> 336          53    4   84
#> 337           0    1   85
#> 338           0    2   85
#> 339          26    3   85
#> 340          26    4   85
#> 341           0    1   86
#> 342           0    2   86
#> 343          54    3   86
#> 344          21    4   86
#> 345           0    1   87
#> 346           0    2   87
#> 347          55    3   87
#> 348           3    4   87
#> 349           0    1   88
#> 350           0    2   88
#> 351          35    3   88
#> 352         106    4   88
#> 353           0    1   89
#> 354           0    2   89
#> 355          35    3   89
#> 356         134    4   89
#> 357           0    1   90
#> 358           0    2   90
#> 359          39    3   90
#> 360          68    4   90
#> 361           0    1   91
#> 362           0    2   91
#> 363          25    3   91
#> 364          91    4   91
#> 365           0    1   92
#> 366           0    2   92
#> 367          40    3   92
#> 368          48    4   92
#> 369           0    1   93
#> 370           0    2   93
#> 371          28    3   93
#> 372          18    4   93
#> 373           0    1   94
#> 374           0    2   94
#> 375          51    3   94
#> 376         130    4   94
#> 377           0    1   95
#> 378           0    2   95
#> 379          31    3   95
#> 380         283    4   95
#> 381           0    1   96
#> 382           0    2   96
#> 383          49    3   96
#> 384          27    4   96
#> 385           0    1   97
#> 386           0    2   97
#> 387          34    3   97
#> 388           0    4   97
#> 389           0    1   98
#> 390           0    2   98
#> 391          30    3   98
#> 392         132    4   98
#> 393           0    1   99
#> 394           0    2   99
#> 395          42    3   99
#> 396         264    4   99
#> 397           0    1  100
#> 398           0    2  100
#> 399          35    3  100
#> 400         204    4  100
#> 401           0    1  101
#> 402           0    2  101
#> 403          47    3  101
#> 404          60    4  101
#> 405           0    1  102
#> 406           0    2  102
#> 407          38    3  102
#> 408           0    4  102
#> 409           0    1  103
#> 410           0    2  103
#> 411          35    3  103
#> 412          25    4  103
#> 413           0    1  104
#> 414           0    2  104
#> 415          30    3  104
#> 416         110    4  104
#> 417           0    1  105
#> 418           0    2  105
#> 419          51    3  105
#> 420          49    4  105
#> 421           0    1  106
#> 422           0    2  106
#> 423          49    3  106
#> 424         104    4  106
#> 425           0    1  107
#> 426           0    2  107
#> 427          51    3  107
#> 428          72    4  107
#> 429           0    1  108
#> 430           0    2  108
#> 431          31    3  108
#> 432           0    4  108
#> 433           0    1  109
#> 434           0    2  109
#> 435          21    3  109
#> 436           0    4  109
#> 437           0    1  110
#> 438           0    2  110
#> 439          26    3  110
#> 440          20    4  110
#> 441           0    1  111
#> 442           0    2  111
#> 443          56    3  111
#> 444           1    4  111
#> 445           0    1  112
#> 446           0    2  112
#> 447          29    3  112
#> 448         115    4  112
#> 449           0    1  113
#> 450           0    2  113
#> 451          25    3  113
#> 452          29    4  113
#> 453           0    1  114
#> 454           0    2  114
#> 455          37    3  114
#> 456         198    4  114
#> 457           0    1  115
#> 458           0    2  115
#> 459          26    3  115
#> 460          22    4  115
#> 461           0    1  116
#> 462           0    2  116
#> 463          44    3  116
#> 464         117    4  116
#> 465           0    1  117
#> 466           0    2  117
#> 467          34    3  117
#> 468         115    4  117
#> 469           0    1  118
#> 470           0    2  118
#> 471          44    3  118
#> 472          67    4  118
#> 473           0    1  119
#> 474           0    2  119
#> 475          24    3  119
#> 476          10    4  119
#> 477           0    1  120
#> 478           0    2  120
#> 479          23    3  120
#> 480          14    4  120
#> 481           0    1  121
#> 482           0    2  121
#> 483          40    3  121
#> 484           1    4  121
#> 485           0    1  122
#> 486           0    2  122
#> 487          31    3  122
#> 488          27    4  122
#> 489           0    1  123
#> 490           0    2  123
#> 491          37    3  123
#> 492           2    4  123
#> 493           0    1  124
#> 494           0    2  124
#> 495          31    3  124
#> 496          10    4  124
#> 497           0    1  125
#> 498           0    2  125
#> 499          37    3  125
#> 500           3    4  125
#> 501           0    1  126
#> 502           0    2  126
#> 503          43    3  126
#> 504          88    4  126
#> 505           0    1  127
#> 506           0    2  127
#> 507          43    3  127
#> 508         457    4  127
#> 509           0    1  128
#> 510           0    2  128
#> 511          44    3  128
#> 512          83    4  128
#> 513           0    1  129
#> 514           0    2  129
#> 515          13    3  129
#> 516         148    4  129
#> 517           0    1  130
#> 518           0    2  130
#> 519          35    3  130
#> 520         101    4  130
#> 521           0    1  131
#> 522           0    2  131
#> 523          50    3  131
#> 524         136    4  131
#> 525           0    1  132
#> 526           0    2  132
#> 527          54    3  132
#> 528         237    4  132
#> 529           0    1  133
#> 530           0    2  133
#> 531          30    3  133
#> 532          51    4  133
#> 533           0    1  134
#> 534           0    2  134
#> 535          30    3  134
#> 536          19    4  134
#> 537           0    1  135
#> 538           0    2  135
#> 539          38    3  135
#> 540          19    4  135
#> 541           0    1  136
#> 542           0    2  136
#> 543          27    3  136
#> 544          43    4  136
#> 545           0    1  137
#> 546           0    2  137
#> 547          34    3  137
#> 548         224    4  137
#> 549           0    1  138
#> 550           0    2  138
#> 551          35    3  138
#> 552          96    4  138
#> 553           0    1  139
#> 554           0    2  139
#> 555          42    3  139
#> 556           1    4  139
#> 557           0    1  140
#> 558           0    2  140
#> 559          41    3  140
#> 560           2    4  140
#> 561           0    1  141
#> 562           0    2  141
#> 563          21    3  141
#> 564         109    4  141
#> 565           0    1  142
#> 566           0    2  142
#> 567          22    3  142
#> 568         210    4  142
#> 569           0    1  143
#> 570           0    2  143
#> 571          50    3  143
#> 572         196    4  143
#> 573           0    1  144
#> 574           0    2  144
#> 575          64    3  144
#> 576         188    4  144
#> 577           0    1  145
#> 578           0    2  145
#> 579          29    3  145
#> 580          22    4  145
#> 581           0    1  146
#> 582           0    2  146
#> 583          26    3  146
#> 584          61    4  146
#> 585           0    1  147
#> 586           0    2  147
#> 587          66    3  147
#> 588          15    4  147
#> 589           0    1  148
#> 590           0    2  148
#> 591          26    3  148
#> 592          69    4  148
#> 593           0    1  149
#> 594           0    2  149
#> 595          62    3  149
#> 596          15    4  149
#> 597           0    1  150
#> 598           0    2  150
#> 599          23    3  150
#> 600           3    4  150
#> 601           0    1  151
#> 602           0    2  151
#> 603          48    3  151
#> 604           9    4  151
#> 605           0    1  152
#> 606           0    2  152
#> 607          31    3  152
#> 608           7    4  152
#> 609           0    1  153
#> 610           0    2  153
#> 611          54    3  153
#> 612         130    4  153
#> 613           0    1  154
#> 614           0    2  154
#> 615          22    3  154
#> 616          21    4  154
#> 617           0    1  155
#> 618           0    2  155
#> 619          41    3  155
#> 620          27    4  155
#> 621           0    1  156
#> 622           0    2  156
#> 623          29    3  156
#> 624          33    4  156
#> 625           0    1  157
#> 626           0    2  157
#> 627          27    3  157
#> 628          66    4  157
#> 629           0    1  158
#> 630           0    2  158
#> 631          35    3  158
#> 632          42    4  158
#> 633           0    1  159
#> 634           0    2  159
#> 635          32    3  159
#> 636         203    4  159
#> 637           0    1  160
#> 638           0    2  160
#> 639          39    3  160
#> 640         207    4  160
#> 641           0    1  161
#> 642           0    2  161
#> 643          51    3  161
#> 644           9    4  161
#> 645           0    1  162
#> 646           0    2  162
#> 647          28    3  162
#> 648          64    4  162
#> 649           0    1  163
#> 650           0    2  163
#> 651          28    3  163
#> 652         127    4  163
#> 653           0    1  164
#> 654           0    2  164
#> 655          32    3  164
#> 656         175    4  164
#> 657           0    1  165
#> 658           0    2  165
#> 659          33    3  165
#> 660           8    4  165
#> 661           0    1  166
#> 662           0    2  166
#> 663          41    3  166
#> 664          22    4  166
#> 665           0    1  167
#> 666           0    2  167
#> 667          66    3  167
#> 668         127    4  167
#> 669           0    1  168
#> 670           0    2  168
#> 671          84    3  168
#> 672           0    4  168
#> 673           0    1  169
#> 674           0    2  169
#> 675          37    3  169
#> 676         120    4  169
#> 677           0    1  170
#> 678           0    2  170
#> 679          31    3  170
#> 680          43    4  170
#> 681           0    1  171
#> 682           0    2  171
#> 683          50    3  171
#> 684         305    4  171
#> 685           0    1  172
#> 686           0    2  172
#> 687          29    3  172
#> 688         272    4  172
#> 689           0    1  173
#> 690           0    2  173
#> 691          58    3  173
#> 692          60    4  173
#> 693           0    1  174
#> 694           0    2  174
#> 695          33    3  174
#> 696          21    4  174
#> 697           0    1  175
#> 698           0    2  175
#> 699          34    3  175
#> 700         163    4  175
#> 701           0    1  176
#> 702           0    2  176
#> 703          28    3  176
#> 704         440    4  176
#> 705           0    1  177
#> 706           0    2  177
#> 707          14    3  177
#> 708         129    4  177
#> 709           0    1  178
#> 710           0    2  178
#> 711          37    3  178
#> 712          11    4  178
#> 713           0    1  179
#> 714           0    2  179
#> 715          19    3  179
#> 716           5    4  179
#> 717           0    1  180
#> 718           0    2  180
#> 719          37    3  180
#> 720          47    4  180
#> 721           0    1  181
#> 722           0    2  181
#> 723          47    3  181
#> 724          36    4  181
#> 725           0    1  182
#> 726           0    2  182
#> 727          19    3  182
#> 728          30    4  182
#> 729           0    1  183
#> 730           0    2  183
#> 731          32    3  183
#> 732          23    4  183
#> 733           0    1  184
#> 734           0    2  184
#> 735          75    3  184
#> 736           1    4  184
#> 737           0    1  185
#> 738           0    2  185
#> 739          26    3  185
#> 740           6    4  185
#> 741           0    1  186
#> 742           0    2  186
#> 743          34    3  186
#> 744           0    4  186
#> 745           0    1  187
#> 746           0    2  187
#> 747          21    3  187
#> 748          88    4  187
#> 749           0    1  188
#> 750           0    2  188
#> 751          27    3  188
#> 752          58    4  188
#> 753           0    1  189
#> 754           0    2  189
#> 755          41    3  189
#> 756          86    4  189
#> 757           0    1  190
#> 758           0    2  190
#> 759          42    3  190
#> 760          50    4  190
#> 761           0    1  191
#> 762           0    2  191
#> 763          38    3  191
#> 764          93    4  191
#> 765           0    1  192
#> 766           0    2  192
#> 767          52    3  192
#> 768          44    4  192
#> 769           0    1  193
#> 770           0    2  193
#> 771          34    3  193
#> 772          29    4  193
#> 773           0    1  194
#> 774           0    2  194
#> 775          34    3  194
#> 776          90    4  194
#> 777           0    1  195
#> 778           0    2  195
#> 779          36    3  195
#> 780          12    4  195
#> 781           0    1  196
#> 782           0    2  196
#> 783          55    3  196
#> 784         501    4  196
#> 785           0    1  197
#> 786           0    2  197
#> 787          36    3  197
#> 788          84    4  197
#> 789           0    1  198
#> 790           0    2  198
#> 791          43    3  198
#> 792         221    4  198
#> 793           0    1  199
#> 794           0    2  199
#> 795          36    3  199
#> 796          56    4  199
#> 797           0    1  200
#> 798           0    2  200
#> 799          39    3  200
#> 800         191    4  200
#> 801           0    1  201
#> 802           0    2  201
#> 803          33    3  201
#> 804          51    4  201
#> 805           0    1  202
#> 806           0    2  202
#> 807          50    3  202
#> 808          39    4  202
#> 809           0    1  203
#> 810           0    2  203
#> 811          34    3  203
#> 812           3    4  203
#> 813           0    1  204
#> 814           0    2  204
#> 815          38    3  204
#> 816          34    4  204
#> 817           0    1  205
#> 818           0    2  205
#> 819          14    3  205
#> 820          22    4  205
#> 821           0    1  206
#> 822           0    2  206
#> 823          34    3  206
#> 824          46    4  206
#> 825           0    1  207
#> 826           0    2  207
#> 827          51    3  207
#> 828          60    4  207
#> 829           0    1  208
#> 830           0    2  208
#> 831          48    3  208
#> 832           0    4  208
#> 833           0    1  209
#> 834           0    2  209
#> 835          32    3  209
#> 836          28    4  209
#> 837           0    1  210
#> 838           0    2  210
#> 839          39    3  210
#> 840           7    4  210
#> 841           0    1  211
#> 842           0    2  211
#> 843          30    3  211
#> 844          40    4  211
#> 845           0    1  212
#> 846           0    2  212
#> 847          43    3  212
#> 848          16    4  212
#> 849           0    1  213
#> 850           0    2  213
#> 851          36    3  213
#> 852          16    4  213
#> 853           0    1  214
#> 854           0    2  214
#> 855          61    3  214
#> 856          37    4  214
#> 857           0    1  215
#> 858           0    2  215
#> 859          31    3  215
#> 860          75    4  215
#> 861           0    1  216
#> 862           0    2  216
#> 863          38    3  216
#> 864          66    4  216
#> 865           0    1  217
#> 866           0    2  217
#> 867          23    3  217
#> 868          37    4  217
#> 869           0    1  218
#> 870           0    2  218
#> 871          37    3  218
#> 872         122    4  218
#> 873           0    1  219
#> 874           0    2  219
#> 875          81    3  219
#> 876           6    4  219
#> 877           0    1  220
#> 878           0    2  220
#> 879          46    3  220
#> 880           4    4  220
#> 881           0    1  221
#> 882           0    2  221
#> 883          36    3  221
#> 884         262    4  221
#> 885           0    1  222
#> 886           0    2  222
#> 887          37    3  222
#> 888         189    4  222
#> 889           0    1  223
#> 890           0    2  223
#> 891          47    3  223
#> 892          20    4  223
#> 893           0    1  224
#> 894           0    2  224
#> 895          41    3  224
#> 896          15    4  224
#> 897           0    1  225
#> 898           0    2  225
#> 899          43    3  225
#> 900         120    4  225
#> 901           0    1  226
#> 902           0    2  226
#> 903          22    3  226
#> 904          50    4  226
#> 905           0    1  227
#> 906           0    2  227
#> 907          22    3  227
#> 908           5    4  227
#> 909           0    1  228
#> 910           0    2  228
#> 911          22    3  228
#> 912           1    4  228
#> 913           0    1  229
#> 914           0    2  229
#> 915          20    3  229
#> 916         240    4  229
#> 917           0    1  230
#> 918           0    2  230
#> 919          46    3  230
#> 920          71    4  230
#> 921           0    1  231
#> 922           0    2  231
#> 923          39    3  231
#> 924           7    4  231
#> 925           0    1  232
#> 926           0    2  232
#> 927          50    3  232
#> 928          13    4  232
#> 929           0    1  233
#> 930           0    2  233
#> 931          45    3  233
#> 932          10    4  233
#> 933           0    1  234
#> 934           0    2  234
#> 935          24    3  234
#> 936          60    4  234
#> 937           0    1  235
#> 938           0    2  235
#> 939          47    3  235
#> 940          35    4  235
#> 941           0    1  236
#> 942           0    2  236
#> 943          51    3  236
#> 944         206    4  236
#> 945           0    1  237
#> 946           0    2  237
#> 947          30    3  237
#> 948           7    4  237
#> 949           0    1  238
#> 950           0    2  238
#> 951          17    3  238
#> 952          34    4  238
#> 953           0    1  239
#> 954           0    2  239
#> 955          43    3  239
#> 956          15    4  239
#> 957           0    1  240
#> 958           0    2  240
#> 959          49    3  240
#> 960         117    4  240
#> 961           0    1  241
#> 962           0    2  241
#> 963          48    3  241
#> 964           5    4  241
#> 965           0    1  242
#> 966           0    2  242
#> 967          47    3  242
#> 968         122    4  242
#> 969           0    1  243
#> 970           0    2  243
#> 971          47    3  243
#> 972          13    4  243
#> 973           0    1  244
#> 974           0    2  244
#> 975          40    3  244
#> 976         147    4  244
#> 977           0    1  245
#> 978           0    2  245
#> 979          46    3  245
#> 980         114    4  245
#> 981           0    1  246
#> 982           0    2  246
#> 983          41    3  246
#> 984          72    4  246
#> 985           0    1  247
#> 986           0    2  247
#> 987          41    3  247
#> 988          13    4  247
#> 989           0    1  248
#> 990           0    2  248
#> 991          37    3  248
#> 992           2    4  248
#> 993           0    1  249
#> 994           0    2  249
#> 995          27    3  249
#> 996         126    4  249
#> 997           0    1  250
#> 998           0    2  250
#> 999          38    3  250
#> 1000         76    4  250
#> 1001          0    1  251
#> 1002          0    2  251
#> 1003         78    3  251
#> 1004          9    4  251
#> 1005          0    1  252
#> 1006          0    2  252
#> 1007         30    3  252
#> 1008         28    4  252
#> 1009          0    1  253
#> 1010          0    2  253
#> 1011         48    3  253
#> 1012         26    4  253
#> 1013          0    1  254
#> 1014          0    2  254
#> 1015         53    3  254
#> 1016         78    4  254
#> 1017          0    1  255
#> 1018          0    2  255
#> 1019         50    3  255
#> 1020         50    4  255
#> 1021          0    1  256
#> 1022          0    2  256
#> 1023         53    3  256
#> 1024         63    4  256
#> 1025          0    1  257
#> 1026          0    2  257
#> 1027         39    3  257
#> 1028         59    4  257
#> 1029          0    1  258
#> 1030          0    2  258
#> 1031         39    3  258
#> 1032        148    4  258
#> 1033          0    1  259
#> 1034          0    2  259
#> 1035         33    3  259
#> 1036        401    4  259
#> 1037          0    1  260
#> 1038          0    2  260
#> 1039         41    3  260
#> 1040        154    4  260
#> 1041          0    1  261
#> 1042          0    2  261
#> 1043         45    3  261
#> 1044         43    4  261
#> 1045          0    1  262
#> 1046          0    2  262
#> 1047         68    3  262
#> 1048        140    4  262
#> 1049          0    1  263
#> 1050          0    2  263
#> 1051         28    3  263
#> 1052        236    4  263
#> 1053          0    1  264
#> 1054          0    2  264
#> 1055         63    3  264
#> 1056         18    4  264
#> 1057          0    1  265
#> 1058          0    2  265
#> 1059         42    3  265
#> 1060        420    4  265
#> 1061          0    1  266
#> 1062          0    2  266
#> 1063         42    3  266
#> 1064        176    4  266
#> 1065          0    1  267
#> 1066          0    2  267
#> 1067         44    3  267
#> 1068        102    4  267
#> 1069          0    1  268
#> 1070          0    2  268
#> 1071         47    3  268
#> 1072        270    4  268
#> 1073          0    1  269
#> 1074          0    2  269
#> 1075         33    3  269
#> 1076        129    4  269
#> 1077          0    1  270
#> 1078          0    2  270
#> 1079         25    3  270
#> 1080         23    4  270
#> 1081          0    1  271
#> 1082          0    2  271
#> 1083         35    3  271
#> 1084         82    4  271
#> 1085          0    1  272
#> 1086          0    2  272
#> 1087         51    3  272
#> 1088          1    4  272
#> 1089          0    1  273
#> 1090          0    2  273
#> 1091         37    3  273
#> 1092        311    4  273
#> 1093          0    1  274
#> 1094          0    2  274
#> 1095         47    3  274
#> 1096          3    4  274
#> 1097          0    1  275
#> 1098          0    2  275
#> 1099         29    3  275
#> 1100         42    4  275
#> 1101          0    1  276
#> 1102          0    2  276
#> 1103         30    3  276
#> 1104          0    4  276
#> 1105          0    1  277
#> 1106          0    2  277
#> 1107         58    3  277
#> 1108         81    4  277
#> 1109          0    1  278
#> 1110          0    2  278
#> 1111         59    3  278
#> 1112        186    4  278
#> 1113          0    1  279
#> 1114          0    2  279
#> 1115         33    3  279
#> 1116          4    4  279
#> 1117          0    1  280
#> 1118          0    2  280
#> 1119         35    3  280
#> 1120         36    4  280
#> 1121          0    1  281
#> 1122          0    2  281
#> 1123         35    3  281
#> 1124         62    4  281
#> 1125          0    1  282
#> 1126          0    2  282
#> 1127         52    3  282
#> 1128         90    4  282
#> 1129          0    1  283
#> 1130          0    2  283
#> 1131         31    3  283
#> 1132         44    4  283
#> 1133          0    1  284
#> 1134          0    2  284
#> 1135         64    3  284
#> 1136         72    4  284
#> 1137          0    1  285
#> 1138          0    2  285
#> 1139         32    3  285
#> 1140         66    4  285
#> 1141          0    1  286
#> 1142          0    2  286
#> 1143         40    3  286
#> 1144        106    4  286
#> 1145          0    1  287
#> 1146          0    2  287
#> 1147         58    3  287
#> 1148        259    4  287
#> 1149          0    1  288
#> 1150          0    2  288
#> 1151         43    3  288
#> 1152          9    4  288
#> 1153          0    1  289
#> 1154          0    2  289
#> 1155         20    3  289
#> 1156         11    4  289
#> 1157          0    1  290
#> 1158          0    2  290
#> 1159         25    3  290
#> 1160         27    4  290
#> 1161          0    1  291
#> 1162          0    2  291
#> 1163         51    3  291
#> 1164        104    4  291
#> 1165          0    1  292
#> 1166          0    2  292
#> 1167         40    3  292
#> 1168         87    4  292
#> 1169          0    1  293
#> 1170          0    2  293
#> 1171         51    3  293
#> 1172        162    4  293
#> 1173          0    1  294
#> 1174          0    2  294
#> 1175         43    3  294
#> 1176        257    4  294
#> 1177          0    1  295
#> 1178          0    2  295
#> 1179         42    3  295
#> 1180        117    4  295
#> 1181          0    1  296
#> 1182          0    2  296
#> 1183         40    3  296
#> 1184          8    4  296
#> 1185          0    1  297
#> 1186          0    2  297
#> 1187         59    3  297
#> 1188         29    4  297
#> 1189          0    1  298
#> 1190          0    2  298
#> 1191         54    3  298
#> 1192         60    4  298
#> 1193          0    1  299
#> 1194          0    2  299
#> 1195         51    3  299
#> 1196        178    4  299
#> 1197          0    1  300
#> 1198          0    2  300
#> 1199         45    3  300
#> 1200         57    4  300
#> 1201          0    1  301
#> 1202          0    2  301
#> 1203         28    3  301
#> 1204        119    4  301
#> 1205          0    1  302
#> 1206          0    2  302
#> 1207         62    3  302
#> 1208         22    4  302
#> 1209          0    1  303
#> 1210          0    2  303
#> 1211         30    3  303
#> 1212          5    4  303
#> 1213          0    1  304
#> 1214          0    2  304
#> 1215         23    3  304
#> 1216         26    4  304
#> 1217          0    1  305
#> 1218          0    2  305
#> 1219         19    3  305
#> 1220        168    4  305
#> 1221          0    1  306
#> 1222          0    2  306
#> 1223         26    3  306
#> 1224         68    4  306
#> 1225          0    1  307
#> 1226          0    2  307
#> 1227         33    3  307
#> 1228        143    4  307
#> 1229          0    1  308
#> 1230          0    2  308
#> 1231         32    3  308
#> 1232         19    4  308
#> 1233          0    1  309
#> 1234          0    2  309
#> 1235         52    3  309
#> 1236        295    4  309
#> 1237          0    1  310
#> 1238          0    2  310
#> 1239         17    3  310
#> 1240         58    4  310
#> 1241          0    1  311
#> 1242          0    2  311
#> 1243         38    3  311
#> 1244         18    4  311
#> 1245          0    1  312
#> 1246          0    2  312
#> 1247         46    3  312
#> 1248        224    4  312
#> 1249          0    1  313
#> 1250          0    2  313
#> 1251         53    3  313
#> 1252        105    4  313
#> 1253          0    1  314
#> 1254          0    2  314
#> 1255         25    3  314
#> 1256         99    4  314
#> 1257          0    1  315
#> 1258          0    2  315
#> 1259         37    3  315
#> 1260         60    4  315
#> 1261          0    1  316
#> 1262          0    2  316
#> 1263         38    3  316
#> 1264        474    4  316
#> 1265          0    1  317
#> 1266          0    2  317
#> 1267         30    3  317
#> 1268          1    4  317
#> 1269          0    1  318
#> 1270          0    2  318
#> 1271         24    3  318
#> 1272         56    4  318
#> 1273          0    1  319
#> 1274          0    2  319
#> 1275         37    3  319
#> 1276        168    4  319
#> 1277          0    1  320
#> 1278          0    2  320
#> 1279         43    3  320
#> 1280         22    4  320
#> 1281          0    1  321
#> 1282          0    2  321
#> 1283         44    3  321
#> 1284         12    4  321
#> 1285          0    1  322
#> 1286          0    2  322
#> 1287         14    3  322
#> 1288         12    4  322
#> 1289          0    1  323
#> 1290          0    2  323
#> 1291         36    3  323
#> 1292         64    4  323
#> 1293          0    1  324
#> 1294          0    2  324
#> 1295         60    3  324
#> 1296         18    4  324
#> 1297          0    1  325
#> 1298          0    2  325
#> 1299         40    3  325
#> 1300         72    4  325
#> 1301          0    1  326
#> 1302          0    2  326
#> 1303         35    3  326
#> 1304        139    4  326
#> 1305          0    1  327
#> 1306          0    2  327
#> 1307         25    3  327
#> 1308         92    4  327
#> 1309          0    1  328
#> 1310          0    2  328
#> 1311         46    3  328
#> 1312         54    4  328
#> 1313          0    1  329
#> 1314          0    2  329
#> 1315         41    3  329
#> 1316         60    4  329
#> 1317          0    1  330
#> 1318          0    2  330
#> 1319         36    3  330
#> 1320        219    4  330
#> 1321          0    1  331
#> 1322          0    2  331
#> 1323         28    3  331
#> 1324        156    4  331
#> 1325          0    1  332
#> 1326          0    2  332
#> 1327         58    3  332
#> 1328         51    4  332
#> 1329          0    1  333
#> 1330          0    2  333
#> 1331         37    3  333
#> 1332        139    4  333
#> 1333          0    1  334
#> 1334          0    2  334
#> 1335         51    3  334
#> 1336         50    4  334
#> 1337          0    1  335
#> 1338          0    2  335
#> 1339         27    3  335
#> 1340        136    4  335
#> 1341          0    1  336
#> 1342          0    2  336
#> 1343         48    3  336
#> 1344          3    4  336
#> 1345          0    1  337
#> 1346          0    2  337
#> 1347         36    3  337
#> 1348        404    4  337
#> 1349          0    1  338
#> 1350          0    2  338
#> 1351         50    3  338
#> 1352         22    4  338
#> 1353          0    1  339
#> 1354          0    2  339
#> 1355         26    3  339
#> 1356         36    4  339
#> 1357          0    1  340
#> 1358          0    2  340
#> 1359         43    3  340
#> 1360          5    4  340
#> 1361          0    1  341
#> 1362          0    2  341
#> 1363         44    3  341
#> 1364         39    4  341
#> 1365          0    1  342
#> 1366          0    2  342
#> 1367         37    3  342
#> 1368         23    4  342
#> 1369          0    1  343
#> 1370          0    2  343
#> 1371         42    3  343
#> 1372         47    4  343
#> 1373          0    1  344
#> 1374          0    2  344
#> 1375         22    3  344
#> 1376         11    4  344
#> 1377          0    1  345
#> 1378          0    2  345
#> 1379         28    3  345
#> 1380        158    4  345
#> 1381          0    1  346
#> 1382          0    2  346
#> 1383         33    3  346
#> 1384         56    4  346
#> 1385          0    1  347
#> 1386          0    2  347
#> 1387         46    3  347
#> 1388         60    4  347
#> 1389          0    1  348
#> 1390          0    2  348
#> 1391         30    3  348
#> 1392          9    4  348
#> 1393          0    1  349
#> 1394          0    2  349
#> 1395         44    3  349
#> 1396         16    4  349
#> 1397          0    1  350
#> 1398          0    2  350
#> 1399         47    3  350
#> 1400         35    4  350
#> 1401          0    1  351
#> 1402          0    2  351
#> 1403         39    3  351
#> 1404         79    4  351
#> 1405          0    1  352
#> 1406          0    2  352
#> 1407         24    3  352
#> 1408         20    4  352
#> 1409          0    1  353
#> 1410          0    2  353
#> 1411         28    3  353
#> 1412         80    4  353
#> 1413          0    1  354
#> 1414          0    2  354
#> 1415         49    3  354
#> 1416         39    4  354
#> 1417          0    1  355
#> 1418          0    2  355
#> 1419         44    3  355
#> 1420         44    4  355
#> 1421          0    1  356
#> 1422          0    2  356
#> 1423         29    3  356
#> 1424          5    4  356
#> 1425          0    1  357
#> 1426          0    2  357
#> 1427         23    3  357
#> 1428        153    4  357
#> 1429          0    1  358
#> 1430          0    2  358
#> 1431         39    3  358
#> 1432         60    4  358
#> 1433          0    1  359
#> 1434          0    2  359
#> 1435         25    3  359
#> 1436        145    4  359
#> 1437          0    1  360
#> 1438          0    2  360
#> 1439         37    3  360
#> 1440         31    4  360
#> 1441          0    1  361
#> 1442          0    2  361
#> 1443         40    3  361
#> 1444         84    4  361
#> 1445          0    1  362
#> 1446          0    2  362
#> 1447         36    3  362
#> 1448          0    4  362
#> 1449          0    1  363
#> 1450          0    2  363
#> 1451         37    3  363
#> 1452         10    4  363
#> 1453          0    1  364
#> 1454          0    2  364
#> 1455         37    3  364
#> 1456        105    4  364
#> 1457          0    1  365
#> 1458          0    2  365
#> 1459         34    3  365
#> 1460         13    4  365
#> 1461          0    1  366
#> 1462          0    2  366
#> 1463         45    3  366
#> 1464        100    4  366
#> 1465          0    1  367
#> 1466          0    2  367
#> 1467         30    3  367
#> 1468         15    4  367
#> 1469          0    1  368
#> 1470          0    2  368
#> 1471         33    3  368
#> 1472        143    4  368
#> 1473          0    1  369
#> 1474          0    2  369
#> 1475         42    3  369
#> 1476        117    4  369
#> 1477          0    1  370
#> 1478          0    2  370
#> 1479         34    3  370
#> 1480         70    4  370
#> 1481          0    1  371
#> 1482          0    2  371
#> 1483         30    3  371
#> 1484         48    4  371
#> 1485          0    1  372
#> 1486          0    2  372
#> 1487         30    3  372
#> 1488         23    4  372
#> 1489          0    1  373
#> 1490          0    2  373
#> 1491         38    3  373
#> 1492        273    4  373
#> 1493          0    1  374
#> 1494          0    2  374
#> 1495         55    3  374
#> 1496         98    4  374
#> 1497          0    1  375
#> 1498          0    2  375
#> 1499         15    3  375
#> 1500         80    4  375
#> 1501          0    1  376
#> 1502          0    2  376
#> 1503         48    3  376
#> 1504         11    4  376
#> 1505          0    1  377
#> 1506          0    2  377
#> 1507         50    3  377
#> 1508        110    4  377
#> 1509          0    1  378
#> 1510          0    2  378
#> 1511         29    3  378
#> 1512        190    4  378
#> 1513          0    1  379
#> 1514          0    2  379
#> 1515         16    3  379
#> 1516          0    4  379
#> 1517          0    1  380
#> 1518          0    2  380
#> 1519         35    3  380
#> 1520        113    4  380
#> 1521          0    1  381
#> 1522          0    2  381
#> 1523         64    3  381
#> 1524         56    4  381
#> 1525          0    1  382
#> 1526          0    2  382
#> 1527         55    3  382
#> 1528        187    4  382
#> 1529          0    1  383
#> 1530          0    2  383
#> 1531         59    3  383
#> 1532         49    4  383
#> 1533          0    1  384
#> 1534          0    2  384
#> 1535         47    3  384
#> 1536          3    4  384
#> 1537          0    1  385
#> 1538          0    2  385
#> 1539         46    3  385
#> 1540         31    4  385
#> 1541          0    1  386
#> 1542          0    2  386
#> 1543         28    3  386
#> 1544         31    4  386
#> 1545          0    1  387
#> 1546          0    2  387
#> 1547         62    3  387
#> 1548         58    4  387
#> 1549          0    1  388
#> 1550          0    2  388
#> 1551         34    3  388
#> 1552         10    4  388
#> 1553          0    1  389
#> 1554          0    2  389
#> 1555         31    3  389
#> 1556         24    4  389
#> 1557          0    1  390
#> 1558          0    2  390
#> 1559         60    3  390
#> 1560        202    4  390
#> 1561          0    1  391
#> 1562          0    2  391
#> 1563         28    3  391
#> 1564        142    4  391
#> 1565          0    1  392
#> 1566          0    2  392
#> 1567         19    3  392
#> 1568         28    4  392
#> 1569          0    1  393
#> 1570          0    2  393
#> 1571         40    3  393
#> 1572          3    4  393
#> 1573          0    1  394
#> 1574          0    2  394
#> 1575         41    3  394
#> 1576         50    4  394
#> 1577          0    1  395
#> 1578          0    2  395
#> 1579         38    3  395
#> 1580        251    4  395
#> 1581          0    1  396
#> 1582          0    2  396
#> 1583         59    3  396
#> 1584        103    4  396
#> 1585          0    1  397
#> 1586          0    2  397
#> 1587         61    3  397
#> 1588          0    4  397
#> 1589          0    1  398
#> 1590          0    2  398
#> 1591         63    3  398
#> 1592        136    4  398
#> 1593          0    1  399
#> 1594          0    2  399
#> 1595         27    3  399
#> 1596         89    4  399
#> 1597          0    1  400
#> 1598          0    2  400
#> 1599         37    3  400
#> 1600          0    4  400
#> 1601          0    1  401
#> 1602          0    2  401
#> 1603         75    3  401
#> 1604         61    4  401
#> 1605          0    1  402
#> 1606          0    2  402
#> 1607         31    3  402
#> 1608         23    4  402
#> 1609          0    1  403
#> 1610          0    2  403
#> 1611         34    3  403
#> 1612          1    4  403
#> 1613          0    1  404
#> 1614          0    2  404
#> 1615         40    3  404
#> 1616         74    4  404
#> 1617          0    1  405
#> 1618          0    2  405
#> 1619         45    3  405
#> 1620          8    4  405
#> 1621          0    1  406
#> 1622          0    2  406
#> 1623         27    3  406
#> 1624        141    4  406
#> 1625          0    1  407
#> 1626          0    2  407
#> 1627         28    3  407
#> 1628         37    4  407
#> 1629          0    1  408
#> 1630          0    2  408
#> 1631         28    3  408
#> 1632         72    4  408
#> 1633          0    1  409
#> 1634          0    2  409
#> 1635         32    3  409
#> 1636          0    4  409
#> 1637          0    1  410
#> 1638          0    2  410
#> 1639         40    3  410
#> 1640        145    4  410
#> 1641          0    1  411
#> 1642          0    2  411
#> 1643         57    3  411
#> 1644         50    4  411
#> 1645          0    1  412
#> 1646          0    2  412
#> 1647         59    3  412
#> 1648          7    4  412
#> 1649          0    1  413
#> 1650          0    2  413
#> 1651         46    3  413
#> 1652         48    4  413
#> 1653          0    1  414
#> 1654          0    2  414
#> 1655         31    3  414
#> 1656         16    4  414
#> 1657          0    1  415
#> 1658          0    2  415
#> 1659         37    3  415
#> 1660        119    4  415
#> 1661          0    1  416
#> 1662          0    2  416
#> 1663         24    3  416
#> 1664         53    4  416
#> 1665          0    1  417
#> 1666          0    2  417
#> 1667         21    3  417
#> 1668         16    4  417
#> 1669          0    1  418
#> 1670          0    2  418
#> 1671         38    3  418
#> 1672         24    4  418
#> 1673          0    1  419
#> 1674          0    2  419
#> 1675         47    3  419
#> 1676         74    4  419
#> 1677          0    1  420
#> 1678          0    2  420
#> 1679         54    3  420
#> 1680         66    4  420
#> 1681          0    1  421
#> 1682          0    2  421
#> 1683         45    3  421
#> 1684         17    4  421
#> 1685          0    1  422
#> 1686          0    2  422
#> 1687         34    3  422
#> 1688         98    4  422
#> 1689          0    1  423
#> 1690          0    2  423
#> 1691         27    3  423
#> 1692         59    4  423
#> 1693          0    1  424
#> 1694          0    2  424
#> 1695         38    3  424
#> 1696         16    4  424
#> 1697          0    1  425
#> 1698          0    2  425
#> 1699         39    3  425
#> 1700         40    4  425
#> 1701          0    1  426
#> 1702          0    2  426
#> 1703         34    3  426
#> 1704         29    4  426
#> 1705          0    1  427
#> 1706          0    2  427
#> 1707         49    3  427
#> 1708        106    4  427
#> 1709          0    1  428
#> 1710          0    2  428
#> 1711         25    3  428
#> 1712         28    4  428
#> 1713          0    1  429
#> 1714          0    2  429
#> 1715         36    3  429
#> 1716         19    4  429
#> 1717          0    1  430
#> 1718          0    2  430
#> 1719         23    3  430
#> 1720          7    4  430
#> 1721          0    1  431
#> 1722          0    2  431
#> 1723         34    3  431
#> 1724         91    4  431
#> 1725          0    1  432
#> 1726          0    2  432
#> 1727         44    3  432
#> 1728        143    4  432
#> 1729          0    1  433
#> 1730          0    2  433
#> 1731         48    3  433
#> 1732         23    4  433
#> 1733          0    1  434
#> 1734          0    2  434
#> 1735         44    3  434
#> 1736          3    4  434
#> 1737          0    1  435
#> 1738          0    2  435
#> 1739         37    3  435
#> 1740         20    4  435
#> 1741          0    1  436
#> 1742          0    2  436
#> 1743         41    3  436
#> 1744         34    4  436
#> 1745          0    1  437
#> 1746          0    2  437
#> 1747         51    3  437
#> 1748          1    4  437
#> 1749          0    1  438
#> 1750          0    2  438
#> 1751         43    3  438
#> 1752        245    4  438
#> 1753          0    1  439
#> 1754          0    2  439
#> 1755         22    3  439
#> 1756          3    4  439
#> 1757          0    1  440
#> 1758          0    2  440
#> 1759         40    3  440
#> 1760         40    4  440
#> 1761          0    1  441
#> 1762          0    2  441
#> 1763         52    3  441
#> 1764         64    4  441
#> 1765          0    1  442
#> 1766          0    2  442
#> 1767         40    3  442
#> 1768        133    4  442
#> 1769          0    1  443
#> 1770          0    2  443
#> 1771         38    3  443
#> 1772        116    4  443
#> 1773          0    1  444
#> 1774          0    2  444
#> 1775         32    3  444
#> 1776        189    4  444
#> 1777          0    1  445
#> 1778          0    2  445
#> 1779         19    3  445
#> 1780         34    4  445
#> 1781          0    1  446
#> 1782          0    2  446
#> 1783         61    3  446
#> 1784         39    4  446
#> 1785          0    1  447
#> 1786          0    2  447
#> 1787         24    3  447
#> 1788         39    4  447
#> 1789          0    1  448
#> 1790          0    2  448
#> 1791         16    3  448
#> 1792         22    4  448
#> 1793          0    1  449
#> 1794          0    2  449
#> 1795         31    3  449
#> 1796          5    4  449
#> 1797          0    1  450
#> 1798          0    2  450
#> 1799         56    3  450
#> 1800         97    4  450
#> 1801          0    1  451
#> 1802          0    2  451
#> 1803         24    3  451
#> 1804         16    4  451
#> 1805          0    1  452
#> 1806          0    2  452
#> 1807         29    3  452
#> 1808         13    4  452
#> 1809          0    1  453
#> 1810          0    2  453
#> 1811         38    3  453
#> 1812         34    4  453
#> 1813          0    1  454
#> 1814          0    2  454
#> 1815         59    3  454
#> 1816         53    4  454
#> 1817          0    1  455
#> 1818          0    2  455
#> 1819         30    3  455
#> 1820         14    4  455
#> 1821          0    1  456
#> 1822          0    2  456
#> 1823         46    3  456
#> 1824          9    4  456
#> 1825          0    1  457
#> 1826          0    2  457
#> 1827         37    3  457
#> 1828         45    4  457
#> 1829          0    1  458
#> 1830          0    2  458
#> 1831         36    3  458
#> 1832        252    4  458
#> 1833          0    1  459
#> 1834          0    2  459
#> 1835         41    3  459
#> 1836         42    4  459
#> 1837          0    1  460
#> 1838          0    2  460
#> 1839         35    3  460
#> 1840         47    4  460
#> 1841          0    1  461
#> 1842          0    2  461
#> 1843         53    3  461
#> 1844         98    4  461
#> 1845          0    1  462
#> 1846          0    2  462
#> 1847         32    3  462
#> 1848          6    4  462
#> 1849          0    1  463
#> 1850          0    2  463
#> 1851         24    3  463
#> 1852        154    4  463
#> 1853          0    1  464
#> 1854          0    2  464
#> 1855         40    3  464
#> 1856        137    4  464
#> 1857          0    1  465
#> 1858          0    2  465
#> 1859         35    3  465
#> 1860         16    4  465
#> 1861          0    1  466
#> 1862          0    2  466
#> 1863         45    3  466
#> 1864         23    4  466
#> 1865          0    1  467
#> 1866          0    2  467
#> 1867         50    3  467
#> 1868         56    4  467
#> 1869          0    1  468
#> 1870          0    2  468
#> 1871         20    3  468
#> 1872        151    4  468
#> 1873          0    1  469
#> 1874          0    2  469
#> 1875         32    3  469
#> 1876         42    4  469
#> 1877          0    1  470
#> 1878          0    2  470
#> 1879         19    3  470
#> 1880         71    4  470
#> 1881          0    1  471
#> 1882          0    2  471
#> 1883         41    3  471
#> 1884         23    4  471
#> 1885          0    1  472
#> 1886          0    2  472
#> 1887         38    3  472
#> 1888          3    4  472
#> 1889          0    1  473
#> 1890          0    2  473
#> 1891         41    3  473
#> 1892         33    4  473
#> 1893          0    1  474
#> 1894          0    2  474
#> 1895         32    3  474
#> 1896        360    4  474
#> 1897          0    1  475
#> 1898          0    2  475
#> 1899         34    3  475
#> 1900         65    4  475
#> 1901          0    1  476
#> 1902          0    2  476
#> 1903         42    3  476
#> 1904          2    4  476
#> 1905          0    1  477
#> 1906          0    2  477
#> 1907         51    3  477
#> 1908        169    4  477
#> 1909          0    1  478
#> 1910          0    2  478
#> 1911         36    3  478
#> 1912         67    4  478
#> 1913          0    1  479
#> 1914          0    2  479
#> 1915         50    3  479
#> 1916         14    4  479
#> 1917          0    1  480
#> 1918          0    2  480
#> 1919         57    3  480
#> 1920         11    4  480
#> 1921          0    1  481
#> 1922          0    2  481
#> 1923         25    3  481
#> 1924         19    4  481
#> 1925          0    1  482
#> 1926          0    2  482
#> 1927         33    3  482
#> 1928         90    4  482
#> 1929          0    1  483
#> 1930          0    2  483
#> 1931         41    3  483
#> 1932         62    4  483
#> 1933          0    1  484
#> 1934          0    2  484
#> 1935         48    3  484
#> 1936         63    4  484
#> 1937          0    1  485
#> 1938          0    2  485
#> 1939         12    3  485
#> 1940         43    4  485
#> 1941          0    1  486
#> 1942          0    2  486
#> 1943         39    3  486
#> 1944        159    4  486
#> 1945          0    1  487
#> 1946          0    2  487
#> 1947         31    3  487
#> 1948         11    4  487
#> 1949          0    1  488
#> 1950          0    2  488
#> 1951         30    3  488
#> 1952         40    4  488
#> 1953          0    1  489
#> 1954          0    2  489
#> 1955         39    3  489
#> 1956        150    4  489
#> 1957          0    1  490
#> 1958          0    2  490
#> 1959         25    3  490
#> 1960        227    4  490
#> 1961          0    1  491
#> 1962          0    2  491
#> 1963         32    3  491
#> 1964          4    4  491
#> 1965          0    1  492
#> 1966          0    2  492
#> 1967         47    3  492
#> 1968        394    4  492
#> 1969          0    1  493
#> 1970          0    2  493
#> 1971         51    3  493
#> 1972        141    4  493
#> 1973          0    1  494
#> 1974          0    2  494
#> 1975         26    3  494
#> 1976        160    4  494
#> 1977          0    1  495
#> 1978          0    2  495
#> 1979         30    3  495
#> 1980          4    4  495
#> 1981          0    1  496
#> 1982          0    2  496
#> 1983         39    3  496
#> 1984         43    4  496
#> 1985          0    1  497
#> 1986          0    2  497
#> 1987         27    3  497
#> 1988         10    4  497
#> 1989          0    1  498
#> 1990          0    2  498
#> 1991         26    3  498
#> 1992         59    4  498
#> 1993          0    1  499
#> 1994          0    2  499
#> 1995         29    3  499
#> 1996        162    4  499
#> 1997          0    1  500
#> 1998          0    2  500
#> 1999         28    3  500
#> 2000         32    4  500