Generate probabilistic nowcast dataframe

This function ingests a list of nowcast matrixes, and generates a long tidy dataframe indexed by time and delay.

Usage

nowcast_matrix_list_to_df(nowcast_matrix_list)

Arguments

nowcast_matrix_list

List of probabilistic nowcast matrices

Value

nowcast_df Dataframe containing observations and probabilistic nowcasts indexed by reference time and delay.

Examples

point_nowcast_matrix <- matrix(
  c(
    80, 50, 25, 10,
    100, 50, 30, 20,
    90, 45, 25, 17.8,
    80, 40, 23.2, 15.9,
    70, 35, 20.3, 19.9
  ),
  nrow = 5,
  byrow = TRUE
)

nowcast_matrix_list <- add_uncertainty(
  point_nowcast_matrix,
  disp = c(8, 1.4, 4),
  n_draws = 10
)

nowcast_df <- nowcast_matrix_list_to_df(nowcast_matrix_list)
nowcast_df
#>     time delay count draw
#> 1      1     1    80    1
#> 2      1     2    50    1
#> 3      1     3    25    1
#> 4      1     4    10    1
#> 5      2     1   100    1
#> 6      2     2    50    1
#> 7      2     3    30    1
#> 8      2     4    20    1
#> 9      3     1    90    1
#> 10     3     2    45    1
#> 11     3     3    25    1
#> 12     3     4    19    1
#> 13     4     1    80    1
#> 14     4     2    40    1
#> 15     4     3    15    1
#> 16     4     4     8    1
#> 17     5     1    70    1
#> 18     5     2    33    1
#> 19     5     3    17    1
#> 20     5     4    31    1
#> 21     1     1    80    2
#> 22     1     2    50    2
#> 23     1     3    25    2
#> 24     1     4    10    2
#> 25     2     1   100    2
#> 26     2     2    50    2
#> 27     2     3    30    2
#> 28     2     4    20    2
#> 29     3     1    90    2
#> 30     3     2    45    2
#> 31     3     3    25    2
#> 32     3     4     9    2
#> 33     4     1    80    2
#> 34     4     2    40    2
#> 35     4     3    22    2
#> 36     4     4    26    2
#> 37     5     1    70    2
#> 38     5     2    43    2
#> 39     5     3     6    2
#> 40     5     4    14    2
#> 41     1     1    80    3
#> 42     1     2    50    3
#> 43     1     3    25    3
#> 44     1     4    10    3
#> 45     2     1   100    3
#> 46     2     2    50    3
#> 47     2     3    30    3
#> 48     2     4    20    3
#> 49     3     1    90    3
#> 50     3     2    45    3
#> 51     3     3    25    3
#> 52     3     4    18    3
#> 53     4     1    80    3
#> 54     4     2    40    3
#> 55     4     3    13    3
#> 56     4     4    29    3
#> 57     5     1    70    3
#> 58     5     2    34    3
#> 59     5     3     1    3
#> 60     5     4    21    3
#> 61     1     1    80    4
#> 62     1     2    50    4
#> 63     1     3    25    4
#> 64     1     4    10    4
#> 65     2     1   100    4
#> 66     2     2    50    4
#> 67     2     3    30    4
#> 68     2     4    20    4
#> 69     3     1    90    4
#> 70     3     2    45    4
#> 71     3     3    25    4
#> 72     3     4     9    4
#> 73     4     1    80    4
#> 74     4     2    40    4
#> 75     4     3    16    4
#> 76     4     4    15    4
#> 77     5     1    70    4
#> 78     5     2    22    4
#> 79     5     3     0    4
#> 80     5     4    20    4
#> 81     1     1    80    5
#> 82     1     2    50    5
#> 83     1     3    25    5
#> 84     1     4    10    5
#> 85     2     1   100    5
#> 86     2     2    50    5
#> 87     2     3    30    5
#> 88     2     4    20    5
#> 89     3     1    90    5
#> 90     3     2    45    5
#> 91     3     3    25    5
#> 92     3     4    16    5
#> 93     4     1    80    5
#> 94     4     2    40    5
#> 95     4     3    26    5
#> 96     4     4    14    5
#> 97     5     1    70    5
#> 98     5     2    29    5
#> 99     5     3     1    5
#> 100    5     4    26    5
#> 101    1     1    80    6
#> 102    1     2    50    6
#> 103    1     3    25    6
#> 104    1     4    10    6
#> 105    2     1   100    6
#> 106    2     2    50    6
#> 107    2     3    30    6
#> 108    2     4    20    6
#> 109    3     1    90    6
#> 110    3     2    45    6
#> 111    3     3    25    6
#> 112    3     4    21    6
#> 113    4     1    80    6
#> 114    4     2    40    6
#> 115    4     3     7    6
#> 116    4     4     8    6
#> 117    5     1    70    6
#> 118    5     2    30    6
#> 119    5     3    29    6
#> 120    5     4    17    6
#> 121    1     1    80    7
#> 122    1     2    50    7
#> 123    1     3    25    7
#> 124    1     4    10    7
#> 125    2     1   100    7
#> 126    2     2    50    7
#> 127    2     3    30    7
#> 128    2     4    20    7
#> 129    3     1    90    7
#> 130    3     2    45    7
#> 131    3     3    25    7
#> 132    3     4    25    7
#> 133    4     1    80    7
#> 134    4     2    40    7
#> 135    4     3    11    7
#> 136    4     4    10    7
#> 137    5     1    70    7
#> 138    5     2    39    7
#> 139    5     3     0    7
#> 140    5     4     1    7
#> 141    1     1    80    8
#> 142    1     2    50    8
#> 143    1     3    25    8
#> 144    1     4    10    8
#> 145    2     1   100    8
#> 146    2     2    50    8
#> 147    2     3    30    8
#> 148    2     4    20    8
#> 149    3     1    90    8
#> 150    3     2    45    8
#> 151    3     3    25    8
#> 152    3     4    19    8
#> 153    4     1    80    8
#> 154    4     2    40    8
#> 155    4     3    68    8
#> 156    4     4    38    8
#> 157    5     1    70    8
#> 158    5     2    22    8
#> 159    5     3    62    8
#> 160    5     4    23    8
#> 161    1     1    80    9
#> 162    1     2    50    9
#> 163    1     3    25    9
#> 164    1     4    10    9
#> 165    2     1   100    9
#> 166    2     2    50    9
#> 167    2     3    30    9
#> 168    2     4    20    9
#> 169    3     1    90    9
#> 170    3     2    45    9
#> 171    3     3    25    9
#> 172    3     4    27    9
#> 173    4     1    80    9
#> 174    4     2    40    9
#> 175    4     3    22    9
#> 176    4     4     1    9
#> 177    5     1    70    9
#> 178    5     2    34    9
#> 179    5     3    27    9
#> 180    5     4    11    9
#> 181    1     1    80   10
#> 182    1     2    50   10
#> 183    1     3    25   10
#> 184    1     4    10   10
#> 185    2     1   100   10
#> 186    2     2    50   10
#> 187    2     3    30   10
#> 188    2     4    20   10
#> 189    3     1    90   10
#> 190    3     2    45   10
#> 191    3     3    25   10
#> 192    3     4    21   10
#> 193    4     1    80   10
#> 194    4     2    40   10
#> 195    4     3     8   10
#> 196    4     4    14   10
#> 197    5     1    70   10
#> 198    5     2    69   10
#> 199    5     3    21   10
#> 200    5     4     4   10