Estimate a delay distribution from a reporting triangle

Provides an estimate of the reporting delay as a function of the delay, based on the reporting triangle and the specified maximum delay and number of reference date observations to be used in the estimation. This point estimate of the delay is computed empirically, using an iterative algorithm starting from the most recent observations. It was modified from the code originally developed by the Karlsruhe Institute of Technology RESPINOW German Hospitalization Nowcasting Hub, Modified from: https://github.com/KITmetricslab/RESPINOW-Hub/blob/7cce3ae2728116e8c8cc0e4ab29074462c24650e/code/baseline/functions.R#L55 #nolint

Usage

get_delay_estimate(
  triangle,
  max_delay = ncol(triangle) - 1,
  n_history = nrow(triangle)
)

Arguments

triangle

matrix of the reporting triangle, with rows representing the time points of reference and columns representing the delays indexed at 0

max_delay

integer indicating the maximum delay to estimate, in units of the delay. The default is to use the whole reporting triangle, ncol(triangle) -1.

n_history

integer indicating the number of reference dates to be used in the estimate of the reporting delay, always starting from the most recent reporting delay. The default is to use the whole reporting triangle, so nrow(triangle)-1

Value

delay_df dataframe of length max_delay with columns delay and pmf, indicating the point estimate of the empirical probability mass on each delay

Examples

library(epinowcast)
#> ! `enw_cache_location` is not set.
#> ℹ Using `tempdir()` at /tmp/Rtmpnw8S8k for the epinowcast model cache location.
#> ℹ Set a specific cache location using `enw_set_cache` to control Stan
#>   recompilation in this R session or across R sessions.
#> ℹ For example: `enw_set_cache(tools::R_user_dir(package = "epinowcast",
#>   "cache"), type = c('session', 'persistent'))`.
#> ℹ See `?enw_set_cache` for details.
#> This message is displayed once per session.
nat_germany_hosp <-
  germany_covid19_hosp[location == "DE"][age_group == "00+"]
nat_germany_hosp <- enw_filter_report_dates(
  nat_germany_hosp,
  latest_date = "2021-10-01"
)
pobs <- enw_preprocess_data(nat_germany_hosp, max_delay = 21)
triangle_raw <- pobs$reporting_triangle[[1]] |>
  dplyr::select(-`.group`, -reference_date) |>
  as.matrix() |>
  unname()
delay_df <- get_delay_estimate(triangle_raw,
  max_delay = 20,
  n_history = 30
)